+++ spec/app/models/basic.rb
@@ -1,0 +1,6 @@
+class Basic
+  include Mongoid::Document
+end
+
+class SubBasic < Basic
+end