+++ spec/app/models/ownable.rb
@@ -1,0 +1,6 @@
+module Ownable
+  extend ActiveSupport::Concern
+  included do
+    belongs_to :user
+  end
+end