+++ spec/unit/mongoid/fields_spec.rb
@@ -263,18 +263,18 @@ describe Mongoid::Fields do
 
         it "calls the handler with the model" do
           handler.expects(:call).with do |model,_,_|
-            model.should eql Person
+            model.should eql User
           end
 
-          Person.field :custom, :option => true
+          User.field :custom, :option => true
         end
 
         it "calls the handler with the field" do
           handler.expects(:call).with do |_,field,_|
-            field.should eql Person.fields["custom"]
+            field.should eql User.fields["custom"]
           end
 
-          Person.field :custom, :option => true
+          User.field :custom, :option => true
         end
 
         it "calls the handler with the option value" do
@@ -282,7 +282,7 @@ describe Mongoid::Fields do
             value.should eql true
           end
 
-          Person.field :custom, :option => true
+          User.field :custom, :option => true
         end
       end
 
@@ -290,7 +290,7 @@ describe Mongoid::Fields do
 
         it "calls the handler" do
           handler.expects(:call)
-          Person.field :custom, :option => nil
+          User.field :custom, :option => nil
         end
       end
 
@@ -299,7 +299,7 @@ describe Mongoid::Fields do
         it "does not call the handler" do
           handler.expects(:call).never
 
-          Person.field :custom
+          User.field :custom
         end
       end
     end
@@ -448,6 +448,35 @@ describe Mongoid::Fields do
 
     it "keeps the options from the old field" do
       new_field.options[:label].should == "id"
+    end
+  end
+
+  context "when sending an include of another module at runtime" do
+
+    before do
+      Basic.send(:include, Ownable)
+    end
+
+    context "when the class is a parent" do
+
+      let(:fields) do
+        Basic.fields
+      end
+
+      it "resets the fields" do
+        fields.keys.should include("user_id")
+      end
+    end
+
+    context "when the class is a subclass" do
+
+      let(:fields) do
+        SubBasic.fields
+      end
+
+      it "resets the fields" do
+        fields.keys.should include("user_id")
+      end
     end
   end
 end